restructure search2 library to top-level bystro library
