+++ bystro/Cargo.toml
@@ -1,16 +1,16 @@
 [package]
-name = "bystro-search"
+name = "bystro"
 version = "0.1.0"
 edition = "2021"
 build = "build.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
-name = "search"
+name = "bystro"
 crate-type = ["cdylib"]
 
 [dependencies]
-arrow = "35.0.0"
+arrow = "40.0.0"
 flate2 = "1.0.25"
 heed = "0.11.0"
 opensearch = "2.0.0"
