+++ bystro/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["maturin>=0.14,<0.15", "setuptools", "wheel", "Cython"]
 build-backend = "maturin"
 
 [project]
-name = "search"
+name = "bystro"
 requires-python = ">=3.10"
 classifiers = [
     "Programming Language :: Rust",
