+++ bystro/python/bystro/__init__.py
@@ -1,0 +1,5 @@
+from .search import *
+
+# __doc__ = bystro.__doc__
+# if hasattr(bystro, "__all__"):
+#     __all__ = bystro.__all__
