+++ bystro/python/bystro/search/__init__.py
@@ -1,0 +1,7 @@
+from .index import *
+from .save import *
+from .utils import *
+
+# __doc__ = search.__doc__
+# if hasattr(search, "__all__"):
+#     __all__ = search.__all__
