+++ bystro/python/bystro/search/index/analysis/amino-synonyms.txt
