+++ bystro/python/bystro/search/index/analysis/codon-map-synonyms.txt
