+++ bystro/python/bystro/search/index/analysis/dbsnp-class-synonyms.txt
