+++ bystro/python/bystro/search/index/analysis/dbsnp-func-synonyms.txt
