+++ bystro/python/bystro/search/index/analysis/type-synonyms.txt
