+++ bystro/python/bystro/search/index/listener.py
@@ -9,9 +9,9 @@ from typing import Optional, Any
 from msgspec import Struct
 from ruamel.yaml import YAML
 
-from search.index.handler import go
-from search.utils.messages import BaseMessage, ProgressPublisher, QueueConf, get_config_file_path, listen
-from search.utils.annotation import AnnotationOutputs
+from bystro.search.index.handler import go
+from bystro.search.utils.messages import BaseMessage, ProgressPublisher, QueueConf, get_config_file_path, listen
+from bystro.search.utils.annotation import AnnotationOutputs
 
 class IndexJobData(BaseMessage, frozen=True):
     inputDir: str
