+++ bystro/python/bystro/search/index/t/test_bystro_indexer.sh
