+++ bystro/python/bystro/search/index/t/test_hg19.mapping.yml
