+++ bystro/python/bystro/search/index/t/test_opensearch_server.sh
