+++ bystro/python/bystro/search/utils/annotation.py
