--- search2/python/search/__init__.py
@@ -1,8 +1,0 @@
-from .search import *
-from .index import *
-from .save import *
-from .utils import *
-
-__doc__ = search.__doc__
-if hasattr(search, "__all__"):
-    __all__ = search.__all__
